requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
pydantic==2.11.7
openai==1.93.0
requests==2.32.4
httpx==0.28.1
tiktoken==0.9.0
python-dotenv==1.0.1
